% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization.R
\name{initialization}
\alias{initialization}
\title{initialize the values used in NMTFOSC}
\usage{
initialization(W,k,l)
}
\arguments{
\item{W}{The matrix to be factorized}

\item{k}{A parameter to specify the row cluster number}

\item{l}{A parameter to specify the column cluster number}
}
\value{
A list with 6 elements, corresponding to the matrices L,C,R,B,Y and the penalty parameter miu
}
\description{
initialize the values which will be updated in NMTFOSC
}
\examples{

W <- simu_data_generation()
init_list <- initialization(W,k=5,l=4)
}
\author{
Xiaoyao Yin
}
