% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu_data_generation.R
\name{simu_data_generation}
\alias{simu_data_generation}
\title{Generate simulation data}
\usage{
simu_data_generation()
}
\value{
The simulated data matrix
}
\description{
To generate the simulation data matrix
}
\examples{

simu_data <- simu_data_generation()
}
\author{
Xiaoyao Yin
}
