% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Standard_Normalization.R
\name{Standard_Normalization}
\alias{Standard_Normalization}
\title{Standard Normalization}
\usage{
Standard_Normalization(x)
}
\arguments{
\item{x}{The data matrix to be normalized}
}
\value{
The normalized matrix
}
\description{
To normalize the data matrix by column
}
\examples{

data1 <- matrix(0,100,100)
data2 <- matrix(0,80,100)
for (i in 1:20)
{
  data1[i,] <- rnorm(100,10,1)
}
for (i in 21:40)
{
  data1[i,] <- rnorm(100,20,1)
}
for (i in 41:60)
{
  data1[i,] <- rnorm(100,30,1)
}
for (i in 61:80)
{
  data1[i,] <- rnorm(100,40,1)
}
for (i in 81:100)
{
  data1[i,] <- rnorm(100,50,1)
}
new_data1 <- Standard_Normalization(data1)
}
\author{
Xiaoyao Yin
}
