% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_B.R
\name{update_B}
\alias{update_B}
\title{Update sub-matrix B}
\usage{
update_B(W,update_L_list)
}
\arguments{
\item{W}{The matrix to be factorized}

\item{update_L_list}{A list containing the updated results in this iteration after running the function update_L}
}
\value{
A list the same as update_L_list with the matrix B updated
}
\description{
Update sub-matrix B
}
\examples{

W <- simu_data_generation()
init_list <- initialization(W,k=5,l=4)
update_L_list <- update_L(W,init_list)
update_B_list <- update_B(W,update_L_list)
}
\author{
Xiaoyao Yin
}
