% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_L.R
\name{update_L}
\alias{update_L}
\title{Update sub-matrix L}
\usage{
update_L(W,init_list)
}
\arguments{
\item{W}{The matrix to be factorized}

\item{init_list}{A list containing the updated results in this iteration}
}
\value{
A list the same as init_list with the matrix L updated
}
\description{
Update sub-matrix L
}
\examples{

W <- simu_data_generation()
init_list <- initialization(W,k=5,l=4)
update_L_list <- update_L(W,init_list)
}
