% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_R.R
\name{update_R}
\alias{update_R}
\title{Update sub-matrix R}
\usage{
update_R(W,update_B_list)
}
\arguments{
\item{W}{The matrix to be factorized}

\item{update_B_list}{A list containing the updated results in this iteration after running the function update_B}
}
\value{
A list the same as update_B_list with the matrix R updated
}
\description{
Update sub-matrix R
}
\examples{

W <- simu_data_generation()
init_list <- initialization(W,k=5,l=4)
update_L_list <- update_L(W,init_list)
update_B_list <- update_B(W,update_L_list)
update_R_list <- update_R(W,update_B_list)
}
